backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Math libraries
numpy==1.26.1
scipy==1.11.3
sympy==1.12
pandas==2.1.2
matplotlib==3.8.1

# Testing
pytest==7.4.3
httpx==0.25.1

# CORS
fastapi-cors==0.0.6

# Background tasks
celery==5.3.4
redis==5.0.1

# Utilities
pillow==10.1.0
pyyaml==6.0.1